Get tRNA precursor sequences

get.tRNAprecursor retrieves tRNA precursor sequences from genomic sequences. The length of 5'- and 3'-overhangs can be specifid individually. The output is checked for validity against the tRNA sequences as reported by tRNAscan.

The chromosomes names of tRNAscan input and genome sequences must be compatible.

get.tRNAprecursor(
  input,
  genome,
  add.5prime = 50L,
  add.3prime = add.5prime,
  trim.intron = FALSE
)

Arguments

input

a compatible GRanges object

genome

a BSgenome object, a DNAStringSet object, a FaFile object or a character vector with a single value referncing a file, which can be coerced to a FaFile object.

add.5prime, add.3prime

the length of overhangs as a single integer value each. (default add.5prime = 50L)

trim.intron

TRUE or FALSE: Should intron sequences be included in the precursor sequences? (default trim.intron = FALSE)

Value

a DNAStringSet object, containing the precursor sequences.

Examples

library(BSgenome.Scerevisiae.UCSC.sacCer3)
#> Loading required package: BSgenome
#> Loading required package: BiocIO
#> Loading required package: rtracklayer
#> 
#> Attaching package: ‘rtracklayer’
#> The following object is masked from ‘package:BiocIO’:
#> 
#>     FileForFormat
file <- system.file("extdata",
                    file = "yeast.tRNAscan",
                    package = "tRNAscanImport")
gr <- tRNAscanImport::import.tRNAscanAsGRanges(file)
genome <- getSeq(BSgenome.Scerevisiae.UCSC.sacCer3)
names(genome) <- c(names(genome)[-17],"chrmt")
get.tRNAprecursor(gr, genome)
#> DNAStringSet object of length 299:
#>       width seq                                             names               
#>   [1]   203 CAATTTGTATATATATACATCT...ATTAAAGTAGCAGTACTTCAAC pre_chrI.tRNA1
#>   [2]   173 AGCTTCTAAGCACTTACCATTC...ATTCGTGAATAGCTGACTGTCA pre_chrI.tRNA2
#>   [3]   214 GTCAGTGTCCAAATAGTTAAAA...ATAATCTACGTAGGAATGAAAG pre_chrI.tRNA3
#>   [4]   182 GTCATACTGACATATCTCATTT...CTCTTCAAAGCATACTCATCTT pre_chrI.tRNA4
#>   [5]   184 GGGTAAAATAGGGTATTTAACT...TTAACTAGAATAATAGGGAAAT pre_chrII.tRNA1
#>   ...   ... ...
#> [295]   174 TAATAAAGAGTAAATATTGTAT...AAATAATATTAATAAATAATTA pre_chrmt.tRNA20
#> [296]   172 TATATATATATATTATATTATA...ATTTATAATAAATATATTATAA pre_chrmt.tRNA21
#> [297]   172 TTATATAATTAATATATTATAT...AATCATTTATAAATTTATTATA pre_chrmt.tRNA24
#> [298]   173 ATAATTAATAAGAGTTTTGGAT...AATATAATAATATTTATTTATT pre_chrmt.tRNA22
#> [299]   173 TATATATATTAGATTTTATAAG...AGTATAATAAAATATATTATAA pre_chrmt.tRNA23
# this produces an error since the seqnames do not match
if (FALSE) {
genome <- BSgenome.Scerevisiae.UCSC.sacCer3
names(genome) <- c(names(genome)[-17],"chrmt")
get.tRNAprecursor(gr, genome)
}
# ... but it can also be fixed
genome <- BSgenome.Scerevisiae.UCSC.sacCer3
seqnames(genome) <- c(seqnames(genome)[-17],"chrmt")
get.tRNAprecursor(gr, genome)
#> DNAStringSet object of length 299:
#>       width seq                                             names               
#>   [1]   203 CAATTTGTATATATATACATCT...ATTAAAGTAGCAGTACTTCAAC pre_chrI.tRNA1
#>   [2]   173 AGCTTCTAAGCACTTACCATTC...ATTCGTGAATAGCTGACTGTCA pre_chrI.tRNA2
#>   [3]   214 GTCAGTGTCCAAATAGTTAAAA...ATAATCTACGTAGGAATGAAAG pre_chrI.tRNA3
#>   [4]   182 GTCATACTGACATATCTCATTT...CTCTTCAAAGCATACTCATCTT pre_chrI.tRNA4
#>   [5]   184 GGGTAAAATAGGGTATTTAACT...TTAACTAGAATAATAGGGAAAT pre_chrII.tRNA1
#>   ...   ... ...
#> [295]   174 TAATAAAGAGTAAATATTGTAT...AAATAATATTAATAAATAATTA pre_chrmt.tRNA20
#> [296]   172 TATATATATATATTATATTATA...ATTTATAATAAATATATTATAA pre_chrmt.tRNA21
#> [297]   172 TTATATAATTAATATATTATAT...AATCATTTATAAATTTATTATA pre_chrmt.tRNA24
#> [298]   173 ATAATTAATAAGAGTTTTGGAT...AATATAATAATATTTATTTATT pre_chrmt.tRNA22
#> [299]   173 TATATATATTAGATTTTATAAG...AGTATAATAAAATATATTATAA pre_chrmt.tRNA23