Importing a tRNAscan output file as a GRanges object

The function import.tRNAscanAsGRanges will import a tRNAscan-SE output file and return the information as a GRanges object. The reported intron sequences are spliced from the result by default, but can also returned as imported.

The function tRNAScan2GFF formats the output of import.tRNAscanAsGRanges to be GFF3 compliant.

tRNAscanID generates a unique tRNA ID, which is like the format used in the SGD annotation

t*AminoAcidSingleLetter*(*Anticodon*)*ChromosomeIdentifier**optionalNumberIfOnTheSameChromosome*

Example: tP(UGG)L or tE(UUC)E1.

import.tRNAscanAsGRanges(input, as.GFF3 = FALSE, trim.intron = TRUE)

tRNAscan2GFF(input)

tRNAscanID(input)

Arguments

input

import.tRNAscanAsGRanges: a tRNAscan-SE input file

tRNAscan2GFF: a compatible GRanges object such as the output of import.tRNAscanAsGRanges

as.GFF3

optional logical for import.tRNAscanAsGRanges: returns a gff3 compatible GRanges object directly. (default: as.GFF3 = FALSE)

trim.intron

optional logical for import.tRNAscanAsGRanges: remove intron sequences. This changes the tRNA length reported. To retrieve the original length fo the tRNA gene, use the width() function on the GRanges object. (default: trim.intron = TRUE)

Value

a GRanges object

References

Chan, Patricia P., and Todd M. Lowe. 2016. “GtRNAdb 2.0: An Expanded Database of Transfer Rna Genes Identified in Complete and Draft Genomes.” Nucleic Acids Research 44 (D1): D184–9. doi:10.1093/nar/gkv1309.

Lowe, T. M., and S. R. Eddy. 1997. “TRNAscan-Se: A Program for Improved Detection of Transfer Rna Genes in Genomic Sequence.” Nucleic Acids Research 25 (5): 955–64.

Examples

gr <- import.tRNAscanAsGRanges(system.file("extdata", 
                               file = "yeast.tRNAscan", 
                               package = "tRNAscanImport"))
gff <- tRNAscan2GFF(gr)
identical(gff,import.tRNAscanAsGRanges(system.file("extdata", 
                               file = "yeast.tRNAscan", 
                               package = "tRNAscanImport"),
                               as.GFF3 = TRUE))
#> [1] TRUE